closecare-backend-test/requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.5
flask-cors==3.0.8
flask-jsonschema==0.1.1
validate-docbr==1.7.0
uwsgi==2.0.19.1
